# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface LoginResult {
  authenticated: boolean;
  displayName: string;
}

export type PurchasePaymentMethod = typeof PurchasePaymentMethod[keyof typeof PurchasePaymentMethod];


export const PurchasePaymentMethod = {
  not_informed: 'not_informed',
  pix: 'pix',
  credit_card: 'credit_card',
  debit_card: 'debit_card',
  boleto: 'boleto',
  cash: 'cash',
  other: 'other',
} as const;

export type PurchaseSource = typeof PurchaseSource[keyof typeof PurchaseSource];


export const PurchaseSource = {
  manual: 'manual',
} as const;

export type PurchaseStatus = typeof PurchaseStatus[keyof typeof PurchaseStatus];


export const PurchaseStatus = {
  pending: 'pending',
  delivered: 'delivered',
  overdue: 'overdue',
} as const;

export interface Purchase {
  id: number;
  purchaseDate: string;
  /** @nullable */
  deliveryDate: string | null;
  supplier: string;
  productName: string;
  recipient: string;
  base: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  totalValue: number;
  paymentMethod: PurchasePaymentMethod;
  source: PurchaseSource;
  status: PurchaseStatus;
  /** @nullable */
  deliveredAt: string | null;
  createdAt: string;
}

export type PurchaseInputPaymentMethod = typeof PurchaseInputPaymentMethod[keyof typeof PurchaseInputPaymentMethod];


export const PurchaseInputPaymentMethod = {
  not_informed: 'not_informed',
  pix: 'pix',
  credit_card: 'credit_card',
  debit_card: 'debit_card',
  boleto: 'boleto',
  cash: 'cash',
  other: 'other',
} as const;

export type PurchaseInputSource = typeof PurchaseInputSource[keyof typeof PurchaseInputSource];


export const PurchaseInputSource = {
  manual: 'manual',
} as const;

export type PurchaseInputStatus = typeof PurchaseInputStatus[keyof typeof PurchaseInputStatus];


export const PurchaseInputStatus = {
  pending: 'pending',
  delivered: 'delivered',
  overdue: 'overdue',
} as const;

export interface PurchaseInput {
  purchaseDate: string;
  /** @nullable */
  deliveryDate?: string | null;
  /** @minLength 1 */
  supplier: string;
  /** @minLength 1 */
  productName: string;
  /** @minLength 1 */
  recipient: string;
  /** @minLength 1 */
  base: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  totalValue: number;
  paymentMethod: PurchaseInputPaymentMethod;
  source: PurchaseInputSource;
  status?: PurchaseInputStatus;
}

export type PurchaseUpdatePaymentMethod = typeof PurchaseUpdatePaymentMethod[keyof typeof PurchaseUpdatePaymentMethod];


export const PurchaseUpdatePaymentMethod = {
  not_informed: 'not_informed',
  pix: 'pix',
  credit_card: 'credit_card',
  debit_card: 'debit_card',
  boleto: 'boleto',
  cash: 'cash',
  other: 'other',
} as const;

export type PurchaseUpdateSource = typeof PurchaseUpdateSource[keyof typeof PurchaseUpdateSource];


export const PurchaseUpdateSource = {
  manual: 'manual',
} as const;

export type PurchaseUpdateStatus = typeof PurchaseUpdateStatus[keyof typeof PurchaseUpdateStatus];


export const PurchaseUpdateStatus = {
  pending: 'pending',
  delivered: 'delivered',
  overdue: 'overdue',
} as const;

export interface PurchaseUpdate {
  purchaseDate?: string;
  /** @nullable */
  deliveryDate?: string | null;
  /** @minLength 1 */
  supplier?: string;
  /** @minLength 1 */
  productName?: string;
  /** @minLength 1 */
  recipient?: string;
  /** @minLength 1 */
  base?: string;
  /** @minimum 1 */
  quantity?: number;
  /** @minimum 0 */
  totalValue?: number;
  paymentMethod?: PurchaseUpdatePaymentMethod;
  source?: PurchaseUpdateSource;
  status?: PurchaseUpdateStatus;
}

export interface Attachment {
  id: number;
  /** @nullable */
  purchaseId: number | null;
  /** @nullable */
  quotationName: string | null;
  purchaseName: string;
  fileName: string;
  mimeType: string;
  fileSize: number;
  createdAt: string;
  downloadUrl: string;
}

export interface AttachmentInput {
  /**
     * @minimum 1
     * @nullable
     */
  purchaseId?: number | null;
  /**
     * @minLength 1
     * @nullable
     */
  quotationName?: string | null;
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  mimeType: string;
  /** @minimum 1 */
  fileSize: number;
  /** @minLength 1 */
  data: string;
}

export interface DashboardSummary {
  total: number;
  pending: number;
  delivered: number;
  overdue: number;
  thisMonth: number;
  recent: Purchase[];
}

export type ExportInputFormat = typeof ExportInputFormat[keyof typeof ExportInputFormat];


export const ExportInputFormat = {
  xlsx: 'xlsx',
  csv: 'csv',
  ods: 'ods',
} as const;

export interface ExportInput {
  format: ExportInputFormat;
  /** @minimum 0 */
  rowCount: number;
}

export type ExportRecordFormat = typeof ExportRecordFormat[keyof typeof ExportRecordFormat];


export const ExportRecordFormat = {
  xlsx: 'xlsx',
  csv: 'csv',
  ods: 'ods',
} as const;

export interface ExportRecord {
  id: number;
  format: ExportRecordFormat;
  rowCount: number;
  createdAt: string;
}