# lib/api-zod/src/generated/types/attachmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttachmentInput {
  /**
     * @minimum 1
     * @nullable
     */
  purchaseId?: number | null;
  /**
     * @minLength 1
     * @nullable
     */
  quotationName?: string | null;
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  mimeType: string;
  /** @minimum 1 */
  fileSize: number;
  /** @minLength 1 */
  data: string;
}