# lib/api-zod/src/generated/types/exportRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExportRecordFormat } from './exportRecordFormat';

export interface ExportRecord {
  id: number;
  format: ExportRecordFormat;
  rowCount: number;
  createdAt: string;
}