# lib/api-zod/src/generated/types/purchase.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PurchasePaymentMethod } from './purchasePaymentMethod';
import type { PurchaseSource } from './purchaseSource';
import type { PurchaseStatus } from './purchaseStatus';

export interface Purchase {
  id: number;
  purchaseDate: Date;
  /** @nullable */
  deliveryDate: Date | null;
  supplier: string;
  productName: string;
  recipient: string;
  base: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  totalValue: number;
  paymentMethod: PurchasePaymentMethod;
  source: PurchaseSource;
  status: PurchaseStatus;
  /** @nullable */
  deliveredAt: string | null;
  createdAt: string;
}