# lib/api-zod/src/generated/types/purchaseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseInputPaymentMethod } from './purchaseInputPaymentMethod';
import type { PurchaseInputSource } from './purchaseInputSource';
import type { PurchaseInputStatus } from './purchaseInputStatus';

export interface PurchaseInput {
  purchaseDate: Date;
  /** @nullable */
  deliveryDate?: Date | null;
  /** @minLength 1 */
  supplier: string;
  /** @minLength 1 */
  productName: string;
  /** @minLength 1 */
  recipient: string;
  /** @minLength 1 */
  base: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  totalValue: number;
  paymentMethod: PurchaseInputPaymentMethod;
  source: PurchaseInputSource;
  status?: PurchaseInputStatus;
}