# lib/api-zod/src/generated/types/purchaseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseUpdatePaymentMethod } from './purchaseUpdatePaymentMethod';
import type { PurchaseUpdateSource } from './purchaseUpdateSource';
import type { PurchaseUpdateStatus } from './purchaseUpdateStatus';

export interface PurchaseUpdate {
  purchaseDate?: Date;
  /** @nullable */
  deliveryDate?: Date | null;
  /** @minLength 1 */
  supplier?: string;
  /** @minLength 1 */
  productName?: string;
  /** @minLength 1 */
  recipient?: string;
  /** @minLength 1 */
  base?: string;
  /** @minimum 1 */
  quantity?: number;
  /** @minimum 0 */
  totalValue?: number;
  paymentMethod?: PurchaseUpdatePaymentMethod;
  source?: PurchaseUpdateSource;
  status?: PurchaseUpdateStatus;
}